Список студентов через typescript
